QTerminal sucks, switching to Terminator
It's almost impossible to use htop in QTerminal. In "default" mode,
backspace does not work, so searching/filtering is almost impossible. In
"linux" mode, backscape works, but then F1/F2/F3 keys don't work.
Additionally, click-dragging when there is a selection tries to drag a
selection, while I want it to start a new selection.

Also trying to add a second shortcut to open the main menu.