Better htop layout
All the CPUs and bars are at the left, all the pure-text entries are on
the right. This keeps related information together and it is easier to
read.

As a downside, on machines with fewer CPUs, there will be more empty
space to the left. Oh, well, I can cope with that.