Even fancier zsh prompt
Now the % is hidden. Unless the user has superuser privileges, then it
is displayed (as #). This leads to a cleaner prompt.

But I also needed a separator for the return code (or exit code) of a
command. The RETURN SYMBOL works well for this, as it is a nice arrow
that makes sense, and it is a play on words due to the name of the
symbol.