Okay, this thing is not working at all.
I don't know why it is not working on the Steam Deck. It should be
working. Maybe we have systemd variables? And maybe the systemd PATH
variable is overwriting the other variable?

By looking at the source code, this could make sense.
https://invent.kde.org/plasma/plasma-workspace/-/blob/master/startkde/startplasma-x11.cpp
And yet it doesn't explain why THIS_DOES_NOT_WORK is not set.